fix calculate_path_finite add ResultData

+++ rusty_fractals_domain/src/lib.rs
@@ -6,4 +6,6 @@ pub mod resolution_multiplier;
 mod domain_pixel_colors;
 mod pixel_states;
 
-const NEIGHBOURS: i32 = 3;
+// How many elements around specific domain element will be investigated for optimization.
+// If there is nothing interesting going on around the specific element, the (new) element will be ignored.
+const NEIGHBOURS: u8 = 3;
